requirements.txt
torch>=1.8.0
torch-geometric>=2.0.0
numpy>=1.19.0
nibabel>=3.2.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
tqdm>=4.60.0
scipy>=1.6.0
pandas>=1.2.0
seaborn>=0.11.0
opencv-python>=4.5.0
scikit-image~=0.25.2
numba
cupy-cuda12x
cuda-bindings
cuda-core
#pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
